+++ components/org.wso2.carbon.identity.api.server.keystore.management/org.wso2.carbon.identity.api.server.keystore.management.v2/src/main/java/org/wso2/carbon/identity/api/server/keystore/management/v2/function/PrivateKeyDataToExternal.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.api.server.keystore.management.v2.function;
+
+import org.wso2.carbon.identity.api.server.keystore.management.v2.model.PrivateKeyDataObject;
+import org.wso2.carbon.security.keystore.service.KeyData;
+
+import java.util.function.Function;
+
+/**
+ * Converts CertData to PrivateKeyData object.
+ */
+public class PrivateKeyDataToExternal implements Function<KeyData, PrivateKeyDataObject> {
+
+
+    @Override
+    public PrivateKeyDataObject apply(KeyData keyData) {
+
+        PrivateKeyDataObject response = new PrivateKeyDataObject();
+
+        response.setAlias(keyData.getAlias());
+        response.setIssurDN(keyData.getIssuerDN());
+        response.setSubjectDN(keyData.getSubjectDN());
+        response.setSerialNumber(Integer.valueOf(String.valueOf(keyData.getSerialNumber())));
+        response.setNotAfter(keyData.getNotAfter());
+        response.setNotBefore(keyData.getNotBefore());
+        response.setVersion(keyData.getVersion());
+        return response;
+    }
+}
